pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "viki-trakt-sync"
version = "0.1.0"
description = "Sync watch history from Viki to Trakt.tv"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Mitchell Currie", email = "mitch@mitchellcurrie.com"}
]
dependencies = [
    "requests>=2.31.0",
    "click>=8.1.7",
    "pytrakt>=4.2.2",
    "python-dotenv>=1.0.0",
    "keyring>=24.3.0",
    "tomli_w>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "responses>=0.24.1",
    "vcrpy>=5.1.0",
    "black>=23.12.1",
    "ruff>=0.1.9",
]

[project.scripts]
viki-trakt = "viki_trakt_sync.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=viki_trakt_sync --cov-report=term-missing"
